+++ src/views/Home.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Redirect } from 'react-router';
 import { IdentityPicker } from '../components/IdentityPicker';
 import { routes } from '../constants/routes';
-import { gaurdian } from './Base';
+import gaurdian from '../utils/KeyGaurdian'
 
 export function Home(): JSX.Element {
     if (gaurdian.hasKey()) {
